+++ inventarioapi/src/main/java/inventario/inventarioapi/model/ItemInventarioDTO.java
@@ -1,6 +1,6 @@
 package inventario.inventarioapi.model;
 
-public record ItemInventarioDTO(Long codInventario, Long codProduto, Double estAtual, Double estLoja) {
+public record ItemInventarioDTO(Long codInventario, Long codProduto, Long estAtual, Long estLoja) {
     public ItemInventarioDTO(ItemInventario object) {
         this(object.getCodigo().getCodInventario(), object.getProduto().getCodigo(), object.getEstAtual(), object.getEstLoja());
     }
